requirements.txt
plotly
numpy
path.py
pandas
scipy